requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.26.0
PyJWT>=2.8.0
